Remove /v0/logs
