+++ api/src/app.ts
@@ -13,7 +13,6 @@ import appRoutes from "./routes/app-routes.js";
 import dependencies from "./routes/dependencies.js";
 import inference from "./routes/inference.js";
 import issues from "./routes/issues.js";
-import logs from "./routes/logs.js";
 import settings from "./routes/settings.js";
 import source from "./routes/source.js";
 import traces from "./routes/traces.js";
@@ -56,7 +55,6 @@ export function createApp(wsConnections?: Set<WebSocket>) {
   );
 
   // All routes are modularized in the ./routes folder
-  app.route("/", logs);
   app.route("/", traces);
   app.route("/", inference);
   app.route("/", source);
